+++ app/Providers/ChiefServiceProvider.php
@@ -237,8 +237,6 @@ class ChiefServiceProvider extends ServiceProvider
 
         Relation::morphMap(['fragmentmodel' => FragmentModel::class]);
 
-        $this->app['view']->addNamespace('chief-site', __DIR__.'/../../resources/views/site');
-
         Blade::directive('fragments', function () {
             return '<?php echo app(\\Thinktomorrow\\Chief\\Fragments\\FragmentsRenderer::class)->render($model instanceof \Thinktomorrow\Chief\Shared\Concerns\Nestable\Tree\NestedNode ? $model->getModel() : $model, get_defined_vars()); ?>';
         });