+++ tests/TestCase.php
@@ -2,6 +2,7 @@
 
 namespace Thinktomorrow\Chief\Tests;
 
+use Livewire\LivewireServiceProvider;
 use Orchestra\Testbench\TestCase as OrchestraTestCase;
 use Thinktomorrow\Chief\App\Providers\ChiefServiceProvider;
 use Thinktomorrow\Chief\Shared\Helpers\Memoize;
@@ -21,6 +22,7 @@ abstract class TestCase extends OrchestraTestCase
     {
         return [
             ChiefServiceProvider::class,
+            LivewireServiceProvider::class,
         ];
     }