+++ CHANGELOG.md
@@ -18,8 +18,10 @@ The Slim script is replaced by a custom livewire solution. This resulted in less
 -   Removed `Thinktomorrow\Chief\Forms\Fields\Media\FileDTO` since now `Thinktomorrow\AssetLibrary\Asset` is passed to file field instead.
 -   Removed `Thinktomorrow\Chief\Managers\Assistants\FileUploadAssistant`, `Thinktomorrow\Chief\Managers\Assistants\SlimImageUploadAssistant` and `Thinktomorrow\Chief\Managers\Assistants\RedactorFileUploadAssistant`. These are no longer used and replaced by the Livewire FileComponent upload logic. The latter one was no longer actively used.
 
+-   Changed: Chief admin toast is refactored as a Chief plugin. This introduces a breaking change in frontend site behaviour. Please refer to its readme file for the frontend setup.
 -   Changed: Updated `<x-chief::nav/>` layout.
 -   Changed: `<x-chief::nav/>` now has a seperate label attribute. This way you can both display a nav title and label.
+-   Removed: `chief-site` view reference. This was only used for the admin toast, which is now its own plugin.
 
 ## 0.7.25 - 2023-05-24